# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 1.0.0
  description: GoalBet Virtual Football Betting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /user/deposit:
    post:
      operationId: deposit
      tags: [user]
      summary: Deposit funds
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DepositRequest"
      responses:
        "200":
          description: Deposit successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionResponse"
...
  /matches:
    get:
      operationId: getMatches
      tags: [matches]
      summary: Get matches
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [upcoming, live, completed]
      responses:
        "200":
          description: Matches
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchListResponse"
...
  /bets:
    post:
      operationId: placeBet
      tags: [bets]
      summary: Place a bet
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PlaceBetRequest"
      responses:
        "201":
          description: Bet placed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetResponse"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /admin/users:
    get:
      operationId: adminGetUsers
      tags: [admin]
      summary: Get all users
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Users list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserListResponse"
  /admin/users/{id}:
    get:
      operationId: adminGetUser
      tags: [admin]
      summary: Get user by ID
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUserResponse"
  /admin/users/{id}/status:
    put:
      operationId: adminUpdateUserStatus
      tags: [admin]
      summary: Ban or suspend user
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateUserStatusRequest"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/users/{id}/balance:
    put:
      operationId: adminAdjustBalance
      tags: [admin]
      summary: Adjust user balance
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdjustBalanceRequest"
      responses:
        "200":
          description: Balance adjusted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/matches:
    post:
      operationId: adminCreateMatch
      tags: [admin]
      summary: Create a match
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateMatchRequest"
      responses:
        "201":
          description: Match created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"
  /admin/matches/{id}/start:
    post:
      operationId: adminStartMatch
      tags: [admin]
      summary: Start match simulation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"
  /admin/matches/{id}/stop:
    post:
      operationId: adminStopMatch
      tags: [admin]
      summary: Stop match simulation
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Match stopped
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"
  /admin/matches/{id}/odds:
    put:
      operationId: adminUpdateOdds
      tags: [admin]
      summary: Update match odds
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateOddsRequest"
      responses:
        "200":
          description: Odds updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"
  /admin/matches/{id}/override:
    post:
      operationId: adminOverrideResult
      tags: [admin]
      summary: Manually override match result
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/OverrideResultRequest"
      responses:
        "200":
          description: Result overridden
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MatchResponse"
  /admin/withdrawals:
    get:
      operationId: adminGetWithdrawals
      tags: [admin]
      summary: Get all withdrawal requests
      security:
        - BearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, approved, rejected]
      responses:
        "200":
          description: Withdrawals
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
  /admin/withdrawals/{id}:
    put:
      operationId: adminProcessWithdrawal
      tags: [admin]
      summary: Approve or reject withdrawal
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProcessWithdrawalRequest"
      responses:
        "200":
          description: Processed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
  /admin/stats:
    get:
      operationId: adminGetStats
      tags: [admin]
      summary: Get platform statistics
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Stats
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminStatsResponse"
  /admin/bets:
    get:
      operationId: adminGetBets
      tags: [admin]
      summary: Get all bets
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
        - name: matchId
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Bets
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/BetListResponse"
  /admin/logs:
    get:
      operationId: adminGetLogs
      tags: [admin]
      summary: Get activity logs
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Logs
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/LogListResponse"
...
  /admin/deposits:
    get:
      operationId: adminGetDeposits
      tags: [admin]
      summary: Get all deposits
      security:
        - BearerAuth: []
      parameters:
        - name: page
          in: query
          schema:
            type: number
        - name: limit
          in: query
          schema:
            type: number
      responses:
        "200":
          description: Deposits
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TransactionListResponse"
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        message:
          type: string
        error:
          type: string
      required: [message]
...
    DepositRequest:
      type: object
      properties:
        amount:
          type: number
        method:
          type: string
      required: [amount]
...
    TransactionResponse:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
          enum: [deposit, withdrawal, bet, winnings, refund, adjustment]
        amount:
          type: number
        fee:
          type: number
        netAmount:
          type: number
        status:
          type: string
          enum: [pending, completed, rejected]
        description:
          type: string
        createdAt:
          type: string
        userId:
          type: string
        username:
          type: string
      required: [id, type, amount, status, createdAt]
...
    MatchListResponse:
      type: object
      properties:
        matches:
          type: array
          items:
            $ref: "#/components/schemas/MatchResponse"
        total:
          type: number
      required: [matches, total]
    PlaceBetRequest:
      type: object
      properties:
        matchId:
          type: string
        outcome:
          type: string
          enum: [home, draw, away]
        amount:
          type: number
      required: [matchId, outcome, amount]
...
    UpdateUserStatusRequest:
      type: object
      properties:
        status:
          type: string
          enum: [active, suspended, banned]
        reason:
          type: string
      required: [status]
...